Move out_mixed_json... functionality to Mixed class
