+++ src/realm/obj.hpp
@@ -56,12 +56,6 @@ namespace _impl {
 class DeepChangeChecker;
 }
 
-enum JSONOutputMode {
-    output_mode_json,       // default / existing implementation for outputting realm to json
-    output_mode_xjson,      // extended json as described in the spec
-    output_mode_xjson_plus, // extended json as described in the spec with additional modifier used for sync
-};
-
 // 'Object' would have been a better name, but it clashes with a class in ObjectStore
 class Obj : public CollectionParent {
 public:
